Module atspi_proxies::accessible

AccessibleProxy

A handle for a remote object implementing the org.a11y.atspi.Accessible interface.

Accessible is the interface which is implemented by all accessible objects.

Structs

AccessibleProxy

AccessibleProxyBlocking