atspi_proxies::application

Struct ApplicationProxyBlocking

pub struct ApplicationProxyBlocking<'p>(/* private fields */);

Application is the interface which is implemented by each accessible application. It is implemented for the root object of an application.

It provides information about the application itself.

Status

A number of methods and properties of this interface have fallen in disuse or are / may be deprecated in the future.

id

set_id

atspi_version

get_locale

toolkit_name and version are still in use.

See the documentation of the individual methods and properties for details.

Implementations

impl<'p> ApplicationProxyBlocking<'p>

pub fn new(conn: &Connection) -> Result<ApplicationProxyBlocking<'p>>

Creates a new proxy with the default service and path.

pub fn builder(conn: &Connection) -> Builder<'p, Self>

Returns a customizable builder for this proxy.

pub fn into_inner(self) -> Proxy<'p>

Consumes self, returning the underlying zbus::Proxy.

pub fn inner(&self) -> &Proxy<'p>

The reference to the underlying zbus::Proxy.

pub fn inner_mut(&mut self) -> &mut Proxy<'p>

The mutable reference to the underlying zbus::Proxy.

pub fn get_locale(&self, lctype: u32) -> Result<String>

Method to retrieve the application’s locale.

Deprecation

This method is likely to be removed in the future.

There is no need to call this method because there is also locale which offers the same functionality at the accessible object level.

See also: Orca issues: “Plans for per-object locale?”

member: “GetLocale”, type: method

pub fn atspi_version(&self) -> Result<String>

retrieves AT-SPI2 version.

Applications are advised to return “2.1” here, thus that is what users should expect.

This was intended to be the version of the atspi interfaces that the application supports, but atspi will probably move to using versioned interface names instead.

member: “AtspiVersion”, type: property

pub fn cached_atspi_version( &self, ) -> Result<Option<<Result<String> as ResultAdapter>::Ok>, <Result<String> as ResultAdapter>::Err>

Get the cached value of the AtspiVersion property, or None if the property is not cached.

pub fn receive_atspi_version_changed( &self, ) -> PropertyIterator<'p, <Result<String> as ResultAdapter>::Ok>

Create a stream for the AtspiVersion property changes. This is a convenient wrapper around zbus::blocking::Proxy::receive_property_changed.

pub fn id(&self) -> Result<i32>

Retrieve numerical id of the application.

The ‘id’ is set an arbitrary numerical id when an application registers with the registry.

When a freshly-started application uses the org.a11y.atspi.Socket’s embed method to register with the accessibility registry, the registry will set a numerical id on the application.

status

The property has fallen in disuse.

As per AT-SPI2-CORE issue #82 it may turn out that this id is not actually used subsequently. This is a remnant of the time when registryd actually had to make up identifiers for each application. With DBus, however, it is the bus that assigns unique names to applications that connect to it.

Applications or toolkits can remember the Id passed when the accessibility registry sets this property, and return it back when the property is read.

member: “Id”, type: property

pub fn cached_id( &self, ) -> Result<Option<<Result<i32> as ResultAdapter>::Ok>, <Result<i32> as ResultAdapter>::Err>

Get the cached value of the Id property, or None if the property is not cached.

pub fn receive_id_changed( &self, ) -> PropertyIterator<'p, <Result<i32> as ResultAdapter>::Ok>

Create a stream for the Id property changes. This is a convenient wrapper around zbus::blocking::Proxy::receive_property_changed.

pub fn set_id(&self, value: i32) -> Result<()>

Set ID of the application.

This method is used by the accessibility registry to set the application’s id.

status

The property has fallen in disuse.

See id for details.

member: “Id”, type: property

pub fn toolkit_name(&self) -> Result<String>

Retrieves the name of the toolkit used to implement the application’s user interface.

member: “ToolkitName”, type: property

pub fn cached_toolkit_name( &self, ) -> Result<Option<<Result<String> as ResultAdapter>::Ok>, <Result<String> as ResultAdapter>::Err>

Get the cached value of the ToolkitName property, or None if the property is not cached.

pub fn receive_toolkit_name_changed( &self, ) -> PropertyIterator<'p, <Result<String> as ResultAdapter>::Ok>

Create a stream for the ToolkitName property changes. This is a convenient wrapper around zbus::blocking::Proxy::receive_property_changed.

pub fn version(&self) -> Result<String>

Returns the version of the toolkit used to implement the application’s user interface.

member: “Version”, type: property

pub fn cached_version( &self, ) -> Result<Option<<Result<String> as ResultAdapter>::Ok>, <Result<String> as ResultAdapter>::Err>

Get the cached value of the Version property, or None if the property is not cached.

pub fn receive_version_changed( &self, ) -> PropertyIterator<'p, <Result<String> as ResultAdapter>::Ok>

Create a stream for the Version property changes. This is a convenient wrapper around zbus::blocking::Proxy::receive_property_changed.