atspi_proxies

Module bus

DBus interface proxies for: org.a11y.Status, org.a11y.Bus

This code was generated by zbus-xmlgen 2.0.1 from DBus introspection data. Source: Interface '/org/a11y/bus' from service 'org.a11y.Bus' on session bus.

You may prefer to adapt it, instead of using it verbatim.

More information can be found in the Writing a client proxy section of the zbus documentation.

This DBus object implements standard DBus interfaces, (org.freedesktop.DBus.*) for which the following zbus proxies can be used:

zbus::fdo::PropertiesProxy

zbus::fdo::IntrospectableProxy

zbus::fdo::PeerProxy

…consequently zbus-xmlgen did not generate code for the above interfaces.

Structs

BusProxy

StatusProxy