atspi_proxies::events

Module event_wrappers

Enums

CacheEvents

All events related to the org.a11y.atspi.Cache interface. Note that these are not telling the client that an item has been added to a cache. It is telling the client “here is a bunch of information to store it in your cache”.

DocumentEvents

Event

Encapsulates the various different accessibility bus signal types.

FocusEvents

KeyboardEvents

MouseEvents

ObjectEvents

TerminalEvents

All events related to the org.a11y.atspi.Event.Terminal interface.

WindowEvents

All events on the org.a11y.atspi.Event.Window interface.