atspi_proxies::events
Module object

Structs

ActiveDescendantChangedEvent

AnnouncementEvent

AttributesChangedEvent

BoundsChangedEvent

ChildrenChangedEvent

A child of an ObjectRef has been added or removed.

ColumnDeletedEvent

A column has been removed from a table.

ColumnInsertedEvent

A column has been added to a table.

ColumnReorderedEvent

A column has been re-ordered within a table.

LinkSelectedEvent

ModelChangedEvent

PropertyChangeEvent

The org.a11y.atspi.Event.Object:PropertyChange event.

RowDeletedEvent

A row has been deleted from a table.

RowInsertedEvent

A row has been added to a table.

RowReorderedEvent

A row has been moved within a table.

SelectionChangedEvent

StateChangedEvent

A state of an object has been modified. A State can be added or removed from any ObjectRef.

TextAttributesChangedEvent

TextBoundsChangedEvent

TextCaretMovedEvent

The caret of the user also known as a cursor (not to be confused with mouse pointer) has changed position.

TextChangedEvent

Text has changed within an ObjectRef.

TextSelectionChangedEvent

VisibleDataChangedEvent

Enums

ObjectEvents

Property