atspi_proxies::events

Module registry

This module contains the events that are emitted by the registry daemon. The events are EventListenerRegisteredEvent and EventListenerDeregisteredEvent.

Modules

socket

This module contains the event that is emitted by the registry daemon’s Socket interface.

Structs

EventListenerDeregisteredEvent

An event that is emitted by the registry daemon, to inform that an event has been deregistered to no longer listen for.

EventListenerRegisteredEvent

An event that is emitted by the regostry daemon to signal that an event has been registered to listen for.

EventListeners

Signal type emitted by EventListenerRegistered and EventListenerDeregistered signals, which belong to the Registry interface, implemented by the registry-daemon.