atspi_proxies::events

Struct EventBody

pub struct EventBody<'a, T> {
    pub kind: T,
    pub detail1: i32,
    pub detail2: i32,
    pub any_data: Value<'a>,
    pub properties: HashMap<UniqueName<'a>, Value<'a>>,
}

A borrowed body for events.

Fields
kind: T

A generic “kind” type, defined by AT-SPI: usually a &'a str, but can be another type like crate::state::State.

detail1: i32

Generic first detail defined by AT-SPI.

detail2: i32

Generic second detail defined by AT-SPI.

any_data: Value<'a>

Generic any_data field defined in AT-SPI. Can contain any type.

properties: HashMap<UniqueName<'a>, Value<'a>>

Map of string to an any type. This is not used for anything, but it is defined by AT-SPI.