atspi_proxies::events

Struct EventBodyOwned

pub struct EventBodyOwned {
    pub kind: String,
    pub detail1: i32,
    pub detail2: i32,
    pub any_data: OwnedValue,
    pub(crate) properties: Properties,
}

AT-SPI2 protocol native event body type.

All of the various signals in the AT-SPI2 protocol share this shape. Most toolkits and implementors emit this type, except for Qt, which has has its own type: EventBodyQtOwned.

Signature (siiva{sv}),

Fields
kind: String

kind variant, used for specifying an event triple “object:state-changed:focused”, the “focus” part of this event is what is contained within the kind.

detail1: i32

Generic detail1 value described by AT-SPI.

detail2: i32

Generic detail2 value described by AT-SPI.

any_data: OwnedValue

Generic any_data value described by AT-SPI. This can be any type.

properties: Properties