Module atspi_proxies::events::terminal

Structs

ApplicationChangedEvent

CharWidthChangedEvent

The width of a terminal emulator has changed sufficiently such that the number of characters able to fit on one visual line has changed.

ColumnCountChangedEvent

The width of a terminal emulator has changed sufficiently such that the number of characters able to fit on one visual line has changed.

LineChangedEvent

A line of text has been changed.

LineCountChangedEvent

The height of a terminal emulator has changed sufficiently such that the number of lines able to fit within the terminal has changed.

Enums

TerminalEvents

All events related to the org.a11y.atspi.Event.Terminal interface.