Trait atspi_proxies::events::HasRegistryEventString

pub trait HasRegistryEventString {
    const REGISTRY_EVENT_STRING: &'static str;
}

A specific trait only to define registry event matches. This is useful for event wrappers like ObjectEvents, which, while it does not have other information required to implement the BusProperties trait, you can indeed add a match rule to the AT-SPI connection to subscribe to all sub events of ObjectEvents.

This trait is not object-safe.

Required Associated Constants

const REGISTRY_EVENT_STRING: &'static str

A registry event string for registering for event receiving via the RegistryProxy. This should be deprecated in favour of composing the string from BusProperties::DBUS_MEMBER and BusProperties::DBUS_INTERFACE.

Object Safety

This trait is not object safe.