Struct atspi_proxies::hyperlink::HyperlinkProxyBlocking

pub struct HyperlinkProxyBlocking<'c>(/* private fields */);

Implementations

impl<'c> HyperlinkProxyBlocking<'c>

pub fn new(conn: &Connection) -> Result<HyperlinkProxyBlocking<'c>>

Creates a new proxy with the default service and path.

pub fn builder(conn: &Connection) -> ProxyBuilder<'c, Self>

Returns a customizable builder for this proxy.

pub fn into_inner(self) -> Proxy<'c>

Consumes self, returning the underlying zbus::Proxy.

pub fn inner(&self) -> &Proxy<'c>

The reference to the underlying zbus::Proxy.

pub fn get_object(&self, i: i32) -> Result<Accessible>

GetObject method

pub fn get_uri(&self, i: i32) -> Result<String>

GetURI method

pub fn is_valid(&self) -> Result<bool>

IsValid method

pub fn end_index(&self) -> Result<i32>

EndIndex property

pub fn cached_end_index( &self ) -> Result<Option<<Result<i32> as ResultAdapter>::Ok>, <Result<i32> as ResultAdapter>::Err>

Get the cached value of the EndIndex property, or None if the property is not cached.

pub fn receive_end_index_changed( &self ) -> PropertyIterator<'c, <Result<i32> as ResultAdapter>::Ok>

Create a stream for the EndIndex property changes. This is a convenient wrapper around zbus::blocking::Proxy::receive_property_changed.

pub fn nanchors(&self) -> Result<i16>

NAnchors property

pub fn cached_nanchors( &self ) -> Result<Option<<Result<i16> as ResultAdapter>::Ok>, <Result<i16> as ResultAdapter>::Err>

Get the cached value of the Nanchors property, or None if the property is not cached.

pub fn receive_nanchors_changed( &self ) -> PropertyIterator<'c, <Result<i16> as ResultAdapter>::Ok>

Create a stream for the Nanchors property changes. This is a convenient wrapper around zbus::blocking::Proxy::receive_property_changed.

pub fn start_index(&self) -> Result<i32>

StartIndex property

pub fn cached_start_index( &self ) -> Result<Option<<Result<i32> as ResultAdapter>::Ok>, <Result<i32> as ResultAdapter>::Err>

Get the cached value of the StartIndex property, or None if the property is not cached.

pub fn receive_start_index_changed( &self ) -> PropertyIterator<'c, <Result<i32> as ResultAdapter>::Ok>

Create a stream for the StartIndex property changes. This is a convenient wrapper around zbus::blocking::Proxy::receive_property_changed.