atspi_proxies::proxy_ext

Struct Proxies

pub struct Proxies<'a> {
    interfaces: InterfaceSet,
    proxy: Proxy<'a>,
}

An object for safe conversion to the related interface proxies.

Fields
interfaces: InterfaceSetproxy: Proxy<'a>
Implementations

impl<'a> Proxies<'a>

pub async fn action(&self) -> Result<ActionProxy<'a>>

Get the Action interface proxy.

Errors

Returns an error if the interface is not available.

pub async fn application(&self) -> Result<ApplicationProxy<'a>>

Get the Application interface proxy.

Errors

Returns an error if the interface is not available.

pub async fn cache(&self) -> Result<CacheProxy<'a>>

Get the Cache interface proxy.

Errors

Returns an error if the interface is not available.

pub async fn collection(&self) -> Result<CollectionProxy<'a>>

Get the Collection interface proxy.

Errors

Returns an error if the interface is not available.

pub async fn component(&self) -> Result<ComponentProxy<'a>>

Get the Component interface proxy.

Errors

Returns an error if the interface is not available.

pub async fn document(&self) -> Result<DocumentProxy<'a>>

Get the Document interface proxy.

Errors

Returns an error if the interface is not available.

pub async fn editable_text(&self) -> Result<EditableTextProxy<'a>>

Get the EditableText interface proxy.

Errors

Returns an error if the interface is not available.

pub async fn hyperlink(&self) -> Result<HyperlinkProxy<'a>>

Get the Hyperlink interface proxy.

Errors

Returns an error if the interface is not available.

pub async fn hypertext(&self) -> Result<HypertextProxy<'a>>

Get the Hypertext interface proxy.

Errors

Returns an error if the interface is not available.

pub async fn image(&self) -> Result<ImageProxy<'a>>

Get the Image interface proxy.

Errors

Returns an error if the interface is not available.

pub async fn selection(&self) -> Result<SelectionProxy<'a>>

Get the Registry interface proxy.

Errors

Returns an error if the interface is not available.

pub async fn table(&self) -> Result<TableProxy<'a>>

Get the Table interface proxy.

Errors

Returns an error if the interface is not available.

pub async fn table_cell(&self) -> Result<TableCellProxy<'a>>

Get the TableCell interface proxy.

Errors

Returns an error if the interface is not available.

pub async fn text(&self) -> Result<TextProxy<'a>>

Get the Text interface proxy.

Errors

Returns an error if the interface is not available.

pub async fn value(&self) -> Result<ValueProxy<'a>>

Get the Value interface proxy.

Errors

Returns an error if the interface is not available.