atspi_proxies::proxy_ext

Trait ProxyExt

pub trait ProxyExt<'a> {
    // Required method
    fn proxies(&self) -> impl Future<Output = Result<Proxies<'a>>>;
}

Acquire the other interface proxies an object may have implemented.

Equip objects with conversions to proxies of the objects’ implemented interfaces by extending AccessibleProxy.

The proxies method returns a Proxies struct.

Required Methods

fn proxies(&self) -> impl Future<Output = Result<Proxies<'a>>>

Get Proxies for the current object.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.