Crate atspi_proxies

Re-exports

pub use error::AtspiError;

pub use zbus;

Modules

accessible

DBus interface proxy for: org.a11y.atspi.Accessible

action

DBus interface proxy for: org.a11y.atspi.Action

application

DBus interface proxy for: org.a11y.atspi.Application

bus

DBus interface proxies for: org.a11y.Status, org.a11y.Bus

cache

DBus interface proxy for: org.a11y.atspi.Cache

collection

DBus interface proxy for: org.a11y.atspi.Collection

component

DBus interface proxy for: org.a11y.atspi.Component

device_event_controller

DBus interface proxy for: org.a11y.atspi.DeviceEventController

device_event_listener

DBus interface proxy for: org.a11y.atspi.DeviceEventListener

document

DBus interface proxy for: org.a11y.atspi.Document

editable_text

DBus interface proxy for: org.a11y.atspi.EditableText

error

events

hyperlink

DBus interface proxy for: org.a11y.atspi.Hyperlink

hypertext

DBus interface proxy for: org.a11y.atspi.Hypertext

image

DBus interface proxy for: org.a11y.atspi.Image

registry

DBus interface proxy for: org.a11y.atspi.Registry

selection

DBus interface proxy for: org.a11y.atspi.Selection

socket

DBus interface proxy for: org.a11y.atspi.Socket

table

DBus interface proxy for: org.a11y.atspi.Table

table_cell

DBus interface proxy for: org.a11y.atspi.TableCell

text

DBus interface proxy for: org.a11y.atspi.Text

value

DBus interface proxy for: org.a11y.atspi.Value

Structs

InterfaceSet

A collection type which encodes the AT-SPI interfaces an accellible object has implemented.

StateSet

The bitflag representation of all states an object may have.

Enums

CoordType

The coordinate type encodes the frame of reference.

Interface

AT-SPI interfaces an accessible object can implement.

State

Used by various interfaces indicating every possible state an crate::accessible::AccessibleProxy object can assume.

Traits

AtspiProxy