atspi_proxies::accessible

Trait ObjectRefExt

pub trait ObjectRefExt {
    // Required method
    fn as_accessible_proxy(
        &self,
        conn: &Connection,
    ) -> impl Future<Output = Result<AccessibleProxy<'_>, Error>> + Send;
}

Required Methods

fn as_accessible_proxy( &self, conn: &Connection, ) -> impl Future<Output = Result<AccessibleProxy<'_>, Error>> + Send

Returns an AccessibleProxy, the handle to the object’s Accessible interface.

Errors

UniqueName or ObjectPath are assumed to be valid because they are obtained from a valid ObjectRef. If the builder is lacking the necessary parameters to build a proxy. See zbus::ProxyBuilder::build. If this method fails, you may want to check the AccessibleProxy default values for missing / invalid parameters.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.