atspi_proxies
Module action

ActionProxy

A handle for a remote object implementing the org.a11y.atspi.Action interface.

The Action interface allows exploring and invoking the actions of a user-actionable UI component.

For example, a button may expose a “click” action - a popup menu may expose an “open” action.

Components which are not “passive” providers of UI information should implement this interface, unless there is a more specialized interface for interaction like org.a11y.atspi.Text or org.a11y.atspi.Value.

Structs

ActionProxy

A handle for a remote object implementing the org.a11y.atspi.Action interface.

ActionProxyBlocking

A handle for a remote object implementing the org.a11y.atspi.Action interface.