Module atspi_proxies::application

ApplicationProxy

A handle for a remote object implementing the org.a11y.atspi.Application interface.

Application is the interface which is implemented by each accessible application. It is implemented for the root object of an application.

It provides information about the application itself.

Status

A number of methods and properties of this interface have fallen in disuse or are / may be deprecated in the future.

id

set_id

atspi_version

get_locale

toolkit_name and version are still in use.

See the documentation of the individual methods and properties for details.

Structs

ApplicationProxy

Application is the interface which is implemented by each accessible application. It is implemented for the root object of an application.

ApplicationProxyBlocking

Application is the interface which is implemented by each accessible application. It is implemented for the root object of an application.