atspi_proxies

Module device_event_controller

DBus interface proxy for: org.a11y.atspi.DeviceEventController

This code was generated by zbus-xmlgen 2.0.1 from DBus introspection data. Source: DeviceEventController.xml.

You may prefer to adapt it, instead of using it verbatim.

More information can be found in the Writing a client proxy section of the zbus documentation.

Structs

DeviceEvent

DeviceEventControllerProxy

EventListenerMode

KeyDefinition

Enums

EventType

KeySynthType