atspi_proxies

Module events

Modules

cache

document

event_body

event_wrappers

focus

keyboard

mouse

object

registry

This module contains the events that are emitted by the registry daemon. The events are EventListenerRegisteredEvent and EventListenerDeregisteredEvent.

terminal

traits

window

Structs

EventBodyBorrowed

EventBodyOwned

AT-SPI2 protocol native event body type.

EventBodyQtBorrowed

EventBodyQtOwned

Event body as used exclusively by ‘Qt’ toolkit.

Enums

CacheEvents

All events related to the org.a11y.atspi.Cache interface. Note that these are not telling the client that an item has been added to a cache. It is telling the client “here is a bunch of information to store it in your cache”.

DocumentEvents

Event

Encapsulates the various different accessibility bus signal types.

EventBody

Common event body that can be either owned or borrowed.

EventListenerEvents

The events that can be emitted by the registry daemon. This enum is used to wrap the events that are emitted by the registry daemon. The events are EventListenerRegisteredEvent and EventListenerDeregisteredEvent.

FocusEvents

KeyboardEvents

MouseEvents

ObjectEvents

TerminalEvents

All events related to the org.a11y.atspi.Event.Terminal interface.

WindowEvents

All events on the org.a11y.atspi.Event.Window interface.

Traits

DBusInterface

The DBus interface name for an event - or a wrapper type. For example, for any event within the “Object” interface, this should be “org.a11y.atspi.Event.Object”.

DBusMatchRule

A static DBus match rule string. This should usually be a string that looks like this: "type='signal',interface='org.a11y.atspi.Event.Object',member='PropertyChange'";

DBusMember

The DBus member for the event. For example, for an crate::events::object::TextChangedEvent this should be "TextChanged"

DBusProperties

A ‘alias’-trait that combines all the DBus related traits.

EventProperties

EventProperties allows access to the internals of an event, specifically:

EventTypeProperties

Describes properties of a specific event type.

MessageConversion

MessageConversionExt

RegistryEventString

A static Registry event string for registering with the RegistryProxy for receiving events.