atspi_proxies::events

Module object

All events which can be received by the org.a11y.atspi.Object interface.

Structs

ActiveDescendantChangedEvent

An event fired when the focus has moved within a tree. The parent: item and descendant (may not be a direct child): descebdant are both referenced for convenience.

AnnouncementEvent

An announcement with a defined text string and an ARIA politeness level.

AttributesChangedEvent

Signal that some attribute of an object (usually styling) has changed. This event does not encode what has changed about the attributes, merely that they have changed.

BoundsChangedEvent

An event triggered when the visual bounds for an item have changed. This usually happens either:

ChildrenChangedEvent

A child of item has been added or removed.

ColumnDeletedEvent

A column has been removed from a table.

ColumnInsertedEvent

A column has been added to a table.

ColumnReorderedEvent

A column has been re-ordered within a table.

LinkSelectedEvent

A link has been selected.

ModelChangedEvent

An event sent when the method of selecting items in a list/set of options changes. Also see: https://docs.gtk.org/gtk4//method.GridView.set_model.html

PropertyChangeEvent

An event representing a property change on UI item item with new value value.

RowDeletedEvent

A row has been deleted from a table.

RowInsertedEvent

A row has been added to a table.

RowReorderedEvent

A row has been moved within a table.

SelectionChangedEvent

The selection of this item has changed. For example: when a selection from a series of checkboxes is changed, this will change the state of the child, and cause a SelectionChangedEvent on the parent.

StateChangedEvent

A state of an object has been modified. A State can be added or removed from any crate::ObjectRef.

TextAttributesChangedEvent

Signal that some attributes about the text (usually styling) have changed. This event does not encode what has changed about the attributes, merely that they have changed.

TextBoundsChangedEvent

The bounds of a piece of text have changed. This event does not specify what the new bounds are; it is only to notify an AT that the bounds have changed. To query information about the new state of the selection, use atspi_proxies::TextProxy’s get_bounded_ranges function.

TextCaretMovedEvent

The caret of the user also known as a cursor (not to be confused with mouse pointer) has changed position.

TextChangedEvent

Text has changed within the UI element item.

TextSelectionChangedEvent

The user’s selection of a piece of text has changed. This event does not specify what the new selection is, nor its indecies; it is only to notify an AT that the selection has changed. To query information about the new state of the selection, use atspi_proxies::TextProxy’s methods.

VisibleDataChangedEvent

A change in whether a particular item is visible or invisible (but still present).

Enums

Property

Any accessibility related property on an crate::ObjectRef. This is used only in the PropertyChangeEvent; this event gets triggered if a role or accessible description of an item changes.