Struct atspi_proxies::events::EventBodyOwned

pub struct EventBodyOwned {
    pub kind: String,
    pub detail1: i32,
    pub detail2: i32,
    pub any_data: OwnedValue,
    pub properties: HashMap<String, OwnedValue>,
}

Standard event body (GTK, egui, etc.) NOTE: Qt has its own signature: EventBodyQT. Signature (siiva{sv}),

Fields
kind: String

kind variant, used for specifying an event triple “object:state-changed:focused”, the “focus” part of this event is what is contained within the kind.

detail1: i32

Generic detail1 value described by AT-SPI.

detail2: i32

Generic detail2 value described by AT-SPI.

any_data: OwnedValue

Generic any_data value described by AT-SPI. This can be any type.

properties: HashMap<String, OwnedValue>

A map of properties. Not in use.