Trait atspi_proxies::events::HasRegistryEventString

pub trait HasRegistryEventString {
    const REGISTRY_EVENT_STRING: &'static str;
}

A specific trait only to define registry event matches.

Required Associated Constants

const REGISTRY_EVENT_STRING: &'static str

A registry event string for registering for event receiving via the RegistryProxy. This should be deprecated in favour of composing the string from GenericEvent::DBUS_MEMBER and GenericEvent::DBUS_INTERFACE.

Object Safety

This trait is not object safe.