atspi_proxies::proxy_ext

Struct Proxies

pub struct Proxies<'a> { /* private fields */ }

An object for safe conversion to the related interface proxies.

Implementations

impl<'a> Proxies<'a>

pub fn action(&mut self) -> Result<&mut ActionProxy<'a>>

Get the Action interface proxy.

Errors

Returns an error if the interface is not available.

pub fn application(&mut self) -> Result<&mut ApplicationProxy<'a>>

Get the Application interface proxy.

Errors

Returns an error if the interface is not available.

pub fn cache(&mut self) -> Result<&mut CacheProxy<'a>>

Get the Cache interface proxy.

Errors

Returns an error if the interface is not available.

pub fn collection(&mut self) -> Result<&mut CollectionProxy<'a>>

Get the Collection interface proxy.

Errors

Returns an error if the interface is not available.

pub fn component(&mut self) -> Result<&mut ComponentProxy<'a>>

Get the Component interface proxy.

Errors

Returns an error if the interface is not available.

pub fn document(&mut self) -> Result<&mut DocumentProxy<'a>>

Get the Document interface proxy.

Errors

Returns an error if the interface is not available.

pub fn editable_text(&mut self) -> Result<&mut EditableTextProxy<'a>>

Get the EditableText interface proxy.

Errors

Returns an error if the interface is not available.

pub fn hyperlink(&mut self) -> Result<&mut HyperlinkProxy<'a>>

Get the Hyperlink interface proxy.

Errors

Returns an error if the interface is not available.

pub fn hypertext(&mut self) -> Result<&mut HypertextProxy<'a>>

Get the Hypertext interface proxy.

Errors

Returns an error if the interface is not available.

pub fn image(&mut self) -> Result<&mut ImageProxy<'a>>

Get the Image interface proxy.

Errors

Returns an error if the interface is not available.

pub fn selection(&mut self) -> Result<&mut SelectionProxy<'a>>

Get the Registry interface proxy.

Errors

Returns an error if the interface is not available.

pub fn table(&mut self) -> Result<&mut TableProxy<'a>>

Get the Table interface proxy.

Errors

Returns an error if the interface is not available.

pub fn table_cell(&mut self) -> Result<&mut TableCellProxy<'a>>

Get the TableCell interface proxy.

Errors

Returns an error if the interface is not available.

pub fn text(&mut self) -> Result<&mut TextProxy<'a>>

Get the Text interface proxy.

Errors

Returns an error if the interface is not available.

pub fn value(&mut self) -> Result<&mut ValueProxy<'a>>

Get the Value interface proxy.

Errors

Returns an error if the interface is not available.