Trait atspi_proxies::proxy_ext::ProxyExt

pub trait ProxyExt<'a> {
    // Required method
    fn proxies(&self) -> impl Future<Output = Result<Proxies<'a>>>;
}

Easily acquire the other interface proxies an object may have.

Equip objects with conversions to proxies of the objects’ further implemented interfaces by extending AccessibleProxy.

The proxies method returns a Proxies struct, which contains lazily loaded proxy accessors.

Lazy initialization and cheap checks

Proxies are lazily initialized, so they are only created when requested. Interface availability is checked before creating the proxy and is a cheap bitop.

Required Methods

fn proxies(&self) -> impl Future<Output = Result<Proxies<'a>>>

Get Proxies for the current object.

Object Safety

This trait is not object safe.