atspi_proxies::proxy_ext

Trait ProxyExt

pub trait ProxyExt<'a> {
    // Required method
    fn proxies(&self) -> impl Future<Output = Result<Proxies<'a>>>;
}

Easily acquire the other interface proxies an object may have.

Equip objects with conversions to proxies of the objects’ further implemented interfaces by extending AccessibleProxy.

The proxies method returns a Proxies struct, which contains lazily loaded proxy accessors.

Lazy initialization and cheap checks

Proxies are lazily initialized, so they are only created when requested. Interface availability is checked before creating the proxy and is a cheap bitop.

Required Methods

fn proxies(&self) -> impl Future<Output = Result<Proxies<'a>>>

Get Proxies for the current object.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.